carechat-microservices/api-gateway-fastapi/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
motor==3.6.0
pymongo==4.10.1
pydantic==2.5.0
pydantic-settings==2.0.3
python-multipart==0.0.6
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-dateutil==2.8.2

# Track2 AI/LLM Dependencies
groq==0.13.0
httpx==0.25.2

# Track2 Audio/Transcription Dependencies  
openai-whisper==20231117
torch==2.1.1
torchaudio==2.1.1
numpy==1.24.3
pydub==0.25.1
ffmpeg-python==0.2.0

# Track2 Translation Dependencies
deep-translator==1.11.4

# Track2 RAG Dependencies
sentence-transformers==2.2.2
scikit-learn==1.3.2
pandas==2.1.3
faiss-cpu==1.7.4

# Track1 Audio Processing Dependencies
librosa==0.10.1
soundfile==0.12.1
